@@ -15,7 +15,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
 // - protoc-gen-go-grpc v1.3.0
-// - protoc             v3.6.1
+// - protoc             v4.25.1
 // source: demo.proto
 
 package oteldemo
