@@ -5,10 +5,11 @@ package kafka
 import (
 	"context"
 	"fmt"
+
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/propagation"
-	semconv "go.opentelemetry.io/otel/semconv/v1.17.0"
+	semconv "go.opentelemetry.io/otel/semconv/v1.21.0"
 	"go.opentelemetry.io/otel/trace"
 
 	"github.com/IBM/sarama"
@@ -21,12 +22,13 @@ type OTelInterceptor struct {
 
 // NewOTelInterceptor processes span for intercepted messages and add some
 // headers with the span data.
-func NewOTelInterceptor() *OTelInterceptor {
+func NewOTelInterceptor(groupID string) *OTelInterceptor {
 	oi := OTelInterceptor{}
 	oi.tracer = otel.Tracer("github.com/open-telemetry/opentelemetry-demo/accountingservice/sarama")
 
 	oi.fixedAttrs = []attribute.KeyValue{
 		semconv.MessagingSystem("kafka"),
+		semconv.MessagingKafkaConsumerGroup(groupID),
 		semconv.NetTransportTCP,
 	}
 	return &oi
@@ -48,8 +50,9 @@ func (oi *OTelInterceptor) OnConsume(msg *sarama.ConsumerMessage) {
 		trace.WithSpanKind(trace.SpanKindConsumer),
 		trace.WithAttributes(oi.fixedAttrs...),
 		trace.WithAttributes(
-			semconv.MessagingDestinationKindTopic,
 			semconv.MessagingDestinationName(msg.Topic),
+			semconv.MessagingKafkaMessageOffset(int(msg.Offset)),
+			semconv.MessagingMessagePayloadSizeBytes(len(msg.Value)),
 			semconv.MessagingOperationReceive,
 			semconv.MessagingKafkaDestinationPartition(int(msg.Partition)),
 		),
