@@ -44,7 +44,7 @@ const CartDropdown = ({ productList, isOpen, onClose }: IProps) => {
           {productList.map(
             ({ quantity, product: { name, picture, id, priceUsd = { nanos: 0, currencyCode: 'USD', units: 0 } } }) => (
               <S.Item key={id} data-cy={CypressFields.CartDropdownItem}>
-                <S.ItemImage src={picture} alt={name} />
+                <S.ItemImage src={"/images/products/" + picture} alt={name} />
                 <S.ItemDetails>
                   <S.ItemName>{name}</S.ItemName>
                   <ProductPrice price={priceUsd} />
